otel-exporters/cd-cli-metrics-exporter/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus
prometheus-client